feat: 前端代码初版

+++ src/main/java/com/macaron/homeschool/common/base/BasePageQuery.java
@@ -26,9 +26,9 @@ import static com.macaron.homeschool.common.constants.MyBatisPageConstants.*;
 @Setter
 public class BasePageQuery {
 
-    private Integer current;
+    private Long current;
 
-    private Integer pageSize;
+    private Long pageSize;
 
     private String sortBy;
